Fixed day toggle test inconsistencies

+++ app/Listeners/UpdateProgressTargets.php
@@ -48,6 +48,8 @@ class UpdateProgressTargets implements ShouldQueueAfterCommit
                 ])->toArray();
             })->toArray();
 
+            if(empty($nextDayTargets)) return;
+
             ExerciseSet::whereIn('day_exercise_id', $nextDayExercises->pluck('id'))->delete();
 
             ExerciseSet::insert($nextDayTargets);
